refactor(frontend): merge Upcoming + Weekly into unified Releases page
Combine the confusingly similar Upcoming and Weekly pages into a single
Releases page with 'My Releases' and 'All Releases' tab views. Old
/upcoming and /weekly URLs redirect to /releases with appropriate view
params. Removes two pages, adds one.
